+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog for NeoFS Node
 
 ### Added
 - Indexes inspection command to neofs-lens (#2882)
+- Add objects sanity checker to neofs-lens (#2506)
 
 ### Fixed
 - Control service's Drop call does not clean metabase (#2822)